+++ app/src/main/java/net/gbs/epp_project/Model/CycleCountDetails.kt
@@ -9,7 +9,7 @@ class CycleCountDetails (
     @SerializedName("itemId"             ) var itemId             : Int?    = null,
     @SerializedName("itemCode"           ) var itemCode           : String? = null,
     @SerializedName("locatorId"          ) var locatorId          : Int?    = null,
-    @SerializedName("qty"                ) var qty                : Int?    = null,
+    @SerializedName("qty"                ) var qty                : Double?    = null,
     @SerializedName("dt"                 ) var dt                 : String? = null,
     @SerializedName("tm"                 ) var tm                 : String? = null,
     @SerializedName("userId"             ) var userId             : String?    = null
