+++ app/src/main/java/net/gbs/epp_project/Ui/Audit/CycleCount/StartCycleCount/ByLocator/StartCycleCountByLocatorViewModel.kt
@@ -49,7 +49,7 @@ class StartCycleCountByLocatorViewModel(private val application: Application,act
     val saveCycleCountStatus = SingleLiveEvent<StatusWithMessage>()
     val saveCycleCountLiveData = SingleLiveEvent<CycleCountHeader>()
 
-    fun saveCycleCount(itemCode:String,locatorCode:String,cycleCountHeaderId:Int,qty:Int,orgCode:String){
+    fun saveCycleCount(itemCode:String,locatorCode:String,cycleCountHeaderId:Int,qty:Double,orgCode:String){
         saveCycleCountStatus.postValue(StatusWithMessage(Status.LOADING))
         job = CoroutineScope(Dispatchers.IO).launch {
             try {
